kill: Extract signal names into signames.h/signames.c
